+++ minecloud_configuration_packages/Terraria/dev/vanilla_configs/advanced_configs/custom-instance-init.ts
@@ -13,7 +13,6 @@ const TERRARIA_SERVER_DOWNLOAD_URL =
 export const CUSTOM_INIT_CONFIG: InitConfig = getCustomInitConfig();
 
 function getCustomInitConfig(): InitConfig {
-
   if (!DEPLOY_LOCAL_SERVER_EXECUTABLE) {
     // Download and setup Terraria server
     return new InitConfig([
@@ -36,7 +35,10 @@ function getCustomInitConfig(): InitConfig {
       InitCommand.shellCommand(`sudo chmod +x TerrariaServer.bin.x86*`, {
         cwd: MINECLOUD_SERVER_DIR
       }),
-      InitFile.fromFileInline(`${MINECLOUD_SERVER_DIR}/serverconfig.txt`,'minecloud_configs/advanced_configs/serverconfig.txt')
+      InitFile.fromFileInline(
+        `${MINECLOUD_SERVER_DIR}/serverconfig.txt`,
+        'minecloud_configs/advanced_configs/serverconfig.txt'
+      )
     ]);
   } else {
     return new InitConfig([