+++ minecloud_configuration_packages/RELEASES.md
@@ -2,10 +2,10 @@
 
 ## Minecraft ([home page](Minecraft/README.md))
 
- - [1.19.4 Vanilla](https://github.com/VeriorPies/MineCloud/raw/main/minecloud_configuration_packages/Minecraft/releases/minecraft_vanilla_1.19.4.zip) (MineCloud 1.2.2)
+ - [1.19.4 Vanilla](https://github.com/VeriorPies/MineCloud/raw/main/minecloud_configuration_packages/Minecraft/releases/minecraft_vanilla_1.19.4.zip) (MineCloud 1.2.3)
 
  ## Factorio ([home page](Factorio/README.md))
- - [1.1.80 Vanilla](https://github.com/VeriorPies/MineCloud/raw/main/minecloud_configuration_packages/Factorio/releases/factorio_vanilla_1.1.80.zip) (MineCloud 1.2.2)
+ - [1.1.80 Vanilla](https://github.com/VeriorPies/MineCloud/raw/main/minecloud_configuration_packages/Factorio/releases/factorio_vanilla_1.1.80.zip) (MineCloud 1.2.3)
 
  ## Terraria ([home page](Terraria/README.md))
- - [1449 Vanilla](https://github.com/VeriorPies/MineCloud/raw/main/minecloud_configuration_packages/Terraria/releases/terraria_vanilla_1449.zip) (MineCloud 1.2.2)
+ - [1449 Vanilla](https://github.com/VeriorPies/MineCloud/raw/main/minecloud_configuration_packages/Terraria/releases/terraria_vanilla_1449.zip) (MineCloud 1.2.3)